requirements.txt
fastapi
mangum
uvicorn
python-dotenv
openai
langchain
gTTS
pyttsx3